center manager passes the tests

+++ src/structure_managers/adaptor_increase_maxorder.hh
@@ -590,7 +590,7 @@ namespace rascal {
     if (traits::MaxOrder-1 == 1) {
       //! Make half neighbour list (strict?)
       //! initialise the neighbourlist
-      for (int i{0}; i < traits::MaxOrder; ++i) {
+      for (size_t i{0}; i < traits::MaxOrder; ++i) {
 	//! this->atom_refs.clear();
 	this->nb_neigh.resize(0);
 	this->offsets.resize(0);
